Add files via upload
The cool part of this code is the implementation of multiple layers of filtering so that one can combine the filters to get the desired features of a cohort. For example, localized disease + T recovery status + with ADT use +Salvage. Furthermore, apply the same filtering to the variant tables can also seamlessly grab the variables saved from the datasets. Note that these variables are not changed based on versions of ADT start and end dates imputation.